refactor: Extract Haversine formula into shared DistanceCalculator service - sonarcloud requirement
