+++ Walkies.API/Controllers/WalkRequestsController.cs
@@ -3,6 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using Walkies.API.Data;
 using Walkies.API.DTOs;
 using Walkies.API.Models;
+using Walkies.API.Services;
 
 namespace Walkies.API.Controllers
 {
@@ -167,7 +168,7 @@ namespace Walkies.API.Controllers
             if (latitude.HasValue && longitude.HasValue && distanceKm.HasValue)
             {
                 walkRequests = walkRequests
-                    .Where(wr => CalculateDistance(
+                    .Where(wr => DistanceCalculator.Calculate(
                         latitude.Value, longitude.Value,
                         wr.Latitude, wr.Longitude) <= distanceKm.Value)
                     .ToList();
@@ -215,29 +216,6 @@ namespace Walkies.API.Controllers
             await _context.SaveChangesAsync();
 
             return NoContent();
-        }
-
-        /// <summary>
-        /// Calulates the distance between two geographic coordinates using the haversine formula.
-        /// </summary>
-        private static double CalculateDistance(double lat1, double lon1, double lat2, double lon2)
-        {
-            const double R = 6371; // Earth radius in kilometers
-            var dLat = ToRadians(lat2 - lat1);
-            var dLon = ToRadians(lon2 - lon1);
-            var a = Math.Sin(dLat / 2) * Math.Sin(dLat / 2) +
-                    Math.Cos(ToRadians(lat1)) * Math.Cos(ToRadians(lat2)) *
-                    Math.Sin(dLon / 2) * Math.Sin(dLon / 2);
-            var c = 2 * Math.Atan2(Math.Sqrt(a), Math.Sqrt(1 - a));
-            return R * c;
-        }
-
-        /// <summary>
-        /// converts degrees to radians.
-        /// </summary>
-        private static double ToRadians(double angle)
-        {
-            return angle * Math.PI / 180;
         }
     }
 }
