+++ Walkies.API/Walkies.API.csproj
@@ -8,7 +8,6 @@
 
   <ItemGroup>
     <Folder Include="Repositories\" />
-    <Folder Include="Services\" />
   </ItemGroup>
 
   <ItemGroup>
